% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{createPlotData}
\alias{createPlotData}
\title{Creates the plot data.}
\usage{
createPlotData(data, byMonth = FALSE)
}
\arguments{
\item{data}{a dataset returned from EunomiaExploration::extractPatients}

\item{byMonth}{Boolean indicating if the time scale is by month or per year (default)}
}
\value{
A data frame with the plot data.
}
\description{
Creates the data for the plotTrend function.
}
