% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{createShinyApp}
\alias{createShinyApp}
\title{Creates a shiny app.}
\usage{
createShinyApp()
}
\value{
A bar plot showing the number of patiens over time, faceted by condition.
}
\description{
Creates a shiny app to visualize condition data.
}
\examples{
#createShinyApp()

}
