% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extractPatients}
\alias{extractPatients}
\title{Extract patient condition data from the Eunomia database}
\usage{
extractPatients(dbConnection)
}
\arguments{
\item{dbConnection}{A database connection created by DatabaseConnector::connect.}
}
\value{
A data frame listing patient conditions.
}
\description{
This function will query the Eunomia database and return data regarding patient
conditions data.
}
\examples{
connectionDetails <- getEunomiaConnectionDetails()

dbConnection <- connect(connectionDetails)

extractPatients(dbConnection)

disconnect(dbConnection)

}
