% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTrend}
\alias{plotTrend}
\title{Creates a trend plot.}
\usage{
plotTrend(data, byMonth = FALSE)
}
\arguments{
\item{data}{a dataset returned from EunomiaExploration::extractPatients}

\item{byMonth}{Boolean indicating if the time scale is by month or per year (default)}
}
\value{
A bar plot showing the number of patiens over time, faceted by condition.
}
\description{
Creates a bar chart showing the number of patients over time, faceted by condition.
}
